+++ test/conftest.py
@@ -1,3 +1,73 @@
+import shutil
+from tempfile import gettempdir
 from .earl import EarlReporter
 
+import pytest
+import os
+from rdflib import ConjunctiveGraph, Dataset
+
+
 pytest_plugins = [EarlReporter.__module__]
+
+
+@pytest.fixture(
+    # scope="function", params=["default", "BerkeleyDB", "SQLiteLSM", "LevelDB", "SQLAlchemy"]
+    scope="function",
+    params=["default"],
+)
+def get_dataset(request):
+    store = request.param
+    tmppath = os.path.join(gettempdir(), f"test_{store.lower()}")
+    graph = Dataset(store=store)
+    graph.open(tmppath, create=True)
+    # tmppath = mkdtemp()
+    if store != "default" and graph.store.is_open():
+        # delete the graph for each test!
+        graph.remove((None, None, None))
+        for c in graph.contexts():
+            c.remove((None, None, None))
+            assert len(c) == 0
+            graph.store.remove_graph(c)
+    # logger.debug(f"Using store {graph.store}")
+    yield graph
+
+    graph.close()
+    if os.path.isdir(tmppath):
+        shutil.rmtree(tmppath)
+    else:
+        try:
+            os.remove(tmppath)
+        except Exception:
+            pass
+
+
+@pytest.fixture(
+    # scope="function", params=["default", "BerkeleyDB", "SQLiteLSM", "LevelDB", "SQLAlchemy"]
+    scope="function",
+    params=["default"],
+)
+def get_conjunctivegraph(request):
+    store = request.param
+    tmppath = os.path.join(gettempdir(), f"test_{store.lower()}")
+    graph = ConjunctiveGraph(store=store)
+    graph.open(tmppath, create=True)
+
+    # tmppath = mkdtemp()
+    if store != "default" and graph.store.is_open():
+        # delete the graph for each test!
+        graph.remove((None, None, None))
+        for c in graph.contexts():
+            c.remove((None, None, None))
+            assert len(c) == 0
+            graph.remove_graph(c)
+    # logger.debug(f"Using store {graph.store}")
+    yield graph
+
+    graph.close()
+    if os.path.isdir(tmppath):
+        shutil.rmtree(tmppath)
+    else:
+        try:
+            os.remove(tmppath)
+        except Exception:
+            pass
